move build_stage_graph where it belongs
